# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dipes & Wipes API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a delivery order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{id}/fulfill-amazon:
    post:
      operationId: fulfillOrderAmazon
      tags: [orders]
      summary: Submit an order to Amazon MCF for fulfillment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FulfillAmazonInput"
      responses:
        "200":
          description: Submitted to Amazon MCF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
        "409":
          description: Order already submitted to fulfillment
        "422":
          description: Unprocessable — missing address or Amazon SKUs
  /orders/{id}/sync-amazon:
    post:
      operationId: syncOrderAmazon
      tags: [orders]
      summary: Sync fulfillment status from Amazon MCF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
        "422":
          description: Order has not been submitted to Amazon MCF
  /orders/one-time:
    post:
      operationId: createOneTimePurchase
      tags: [orders]
      summary: Create a one-time purchase order (not tied to a subscription)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OneTimePurchaseInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
        "404":
          description: Product not found
  /orders/validate-amazon:
    post:
      operationId: validateAmazonCart
      tags: [orders]
      summary: Validate a list of subscription items against the active Amazon MCF SKU whitelist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateAmazonCartInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateAmazonCartResult"
  /orders/process-pending:
    post:
      operationId: processPendingOrders
      tags: [orders]
      summary: Batch-submit all pending orders to Amazon MCF (background job trigger)
      description: |
        Finds all orders with status="pending" (and no fulfillmentId) and attempts
        Amazon MCF auto-fulfillment for each. Pass ?retry=true to also retry orders
        that previously failed with status="fulfillment_error".
      parameters:
        - in: query
          name: retry
          schema:
            type: boolean
          description: When true, also retry orders with status="fulfillment_error"
      responses:
        "200":
          description: Summary of the batch run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessPendingResult"
...
components:
  schemas:
...
    OneTimePurchaseInput:
      type: object
      required: [productId, quantity, scheduledDate, shippingName, shippingAddress1, shippingCity, shippingState, shippingZip]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        diaperSize:
          type: string
          description: Required when purchasing diapers
        scheduledDate:
          type: string
          format: date
        shippingName:
          type: string
        shippingAddress1:
          type: string
        shippingCity:
          type: string
        shippingState:
          type: string
        shippingZip:
          type: string
        shippingCountry:
          type: string
    Order:
      type: object
      required: [id, status, scheduledDate, createdAt]
      properties:
        id:
          type: integer
        subscriptionId:
          type: ["integer", "null"]
        babyId:
          type: ["integer", "null"]
        babyName:
          type: ["string", "null"]
        status:
          type: string
        scheduledDate:
          type: string
          format: date
        diaperSize:
          type: ["string", "null"]
        totalCents:
          type: ["integer", "null"]
        trackingNumber:
          type: ["string", "null"]
        carrier:
          type: ["string", "null"]
          description: Shipping carrier (e.g. USPS, UPS, FedEx)
        fulfillmentProvider:
          type: ["string", "null"]
        fulfillmentId:
          type: ["string", "null"]
        fulfillmentStatus:
          type: ["string", "null"]
        fulfillmentErrorMessage:
          type: ["string", "null"]
          description: Human-readable reason for a failed auto-fulfillment attempt. Set when status is "fulfillment_error".
        shippingName:
          type: ["string", "null"]
        shippingAddress1:
          type: ["string", "null"]
        shippingCity:
          type: ["string", "null"]
        shippingState:
          type: ["string", "null"]
        shippingZip:
          type: ["string", "null"]
        shippingCountry:
          type: ["string", "null"]
        createdAt:
          type: string
...
    FulfillAmazonInput:
      type: object
      properties:
        shippingName:
          type: string
        shippingAddress1:
          type: string
        shippingCity:
          type: string
        shippingState:
          type: string
        shippingZip:
          type: string
        shippingCountry:
          type: string
        shippingEmail:
          type: string
        shippingSpeed:
          type: string
          enum: [Standard, Expedited, Priority]
          default: Standard
    ValidateAmazonCartInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
                minimum: 1
    ValidateAmazonCartResult:
      type: object
      required: [valid, validItems, invalidItems]
      properties:
        valid:
          type: boolean
        validItems:
          type: array
          items:
            type: object
            required: [productId, name, quantity]
            properties:
              productId:
                type: integer
              name:
                type: string
              quantity:
                type: integer
        invalidItems:
          type: array
          items:
            type: object
            required: [productId, reason]
            properties:
              productId:
                type: integer
              reason:
                type: string
    ProcessPendingResult:
      type: object
      required: [processed, succeeded, skipped, failed, results]
      properties:
        processed:
          type: integer
        succeeded:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            type: object
            required: [orderId, success, skipped]
            properties:
              orderId:
                type: integer
              success:
                type: boolean
              skipped:
                type: boolean
              skipReason:
                type: string
              error:
                type: string
...